+++ rapid.h
@@ -7,8 +7,8 @@ class Rapid : public MagStim
 public:
     Rapid(QString serialConnection, int superRapid, QString unlockCode, int voltage, std::tuple<int, int, int> version);
     Rapid(QString serialConnection);
-    float getRapidMinWaitTime(int power, int nPulses, double frequency);
-    float getRapidMaxOnTime(int power, double frequency);
+    float getRapidMinWaitTime(int power, int nPulses, float frequency);
+    float getRapidMaxOnTime(int power, float frequency);
     float getRapidMaxContinuousOperationsFrequency(int power);
 
 private:
